DEFAULT_PACKET_LENGTH: move from PHC to HeaderSpace
HeaderSpace is much closer to being a 'core' datamodel class; it does not have
gross dependencies like on specifiers.

This makes some downstream work much easier.

commit-id:52cd4284
