+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/HeaderSpace.java
@@ -22,6 +22,9 @@ import javax.annotation.Nullable;
 
 public class HeaderSpace implements Serializable, Comparable<HeaderSpace> {
 
+  /** default packet length */
+  public static final int DEFAULT_PACKET_LENGTH = 512;
+
   private static <C extends Collection<?>> C nullIfEmpty(C collection) {
     return collection == null ? null : collection.isEmpty() ? null : collection;
   }
